+++ lib/src/core.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 import 'dart:io';
+
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:pedantic/pedantic.dart';
@@ -54,7 +55,8 @@ class RateMyApp {
   /// Initializes the plugin (loads base launch date, app launches and whether the dialog should not be opened again).
   Future<void> init() async {
     SharedPreferences preferences = await SharedPreferences.getInstance();
-    conditions.forEach((condition) => condition.readFromPreferences(preferences, preferencesPrefix));
+    conditions.forEach((condition) =>
+        condition.readFromPreferences(preferences, preferencesPrefix));
     await callEvent(RateMyAppEventType.initialized);
   }
 
@@ -98,11 +100,13 @@ class RateMyApp {
   }
 
   /// Returns whether native review dialog is supported.
-  Future<bool?> get isNativeReviewDialogSupported => _channel.invokeMethod<bool>('isNativeDialogSupported');
+  Future<bool?> get isNativeReviewDialogSupported =>
+      _channel.invokeMethod<bool>('isNativeDialogSupported');
 
   /// Launches the native review dialog.
   /// You should check for [isNativeReviewDialogSupported] before running the method.
-  Future<void> launchNativeReviewDialog() => _channel.invokeMethod('launchNativeReviewDialog');
+  Future<void> launchNativeReviewDialog() =>
+      _channel.invokeMethod('launchNativeReviewDialog');
 
   /// Shows the rate dialog.
   Future<void> showRateDialog(
@@ -120,21 +124,23 @@ class RateMyApp {
     VoidCallback? onDismissed,
     bool? barrierDismissible,
     String? barrierLabel,
-    DialogTransition dialogTransition = const DialogTransition.def(),
+    DialogTransition dialogTransition = const DialogTransition(),
   }) async {
     ignoreNativeDialog ??= Platform.isAndroid;
-    if (!ignoreNativeDialog && ((await isNativeReviewDialogSupported) ?? false)) {
+    if (!ignoreNativeDialog &&
+        ((await isNativeReviewDialogSupported) ?? false)) {
       unawaited(callEvent(RateMyAppEventType.iOSRequestReview));
       await launchNativeReviewDialog();
       return;
     }
 
-    var rateMyAppDialog = RateMyAppDialog(
+    RateMyAppDialog rateMyAppDialog = RateMyAppDialog(
       this,
       title: title ?? 'Rate this app',
       message: message ??
           'If you like this app, please take a little bit of your time to review it !\nIt really helps us and it shouldn\'t take you more than one minute.',
-      contentBuilder: contentBuilder ?? ((context, defaultContent) => defaultContent),
+      contentBuilder:
+          contentBuilder ?? ((context, defaultContent) => defaultContent),
       actionsBuilder: actionsBuilder,
       rateButton: rateButton ?? 'RATE',
       noButton: noButton ?? 'NO THANKS',
@@ -146,8 +152,11 @@ class RateMyApp {
     unawaited(callEvent(RateMyAppEventType.dialogOpen));
 
     // Using showDialog() when TransitionType.none to get rid of the default fading animation of showGeneralDialog()
-    RateMyAppDialogButton? clickedButton = await (dialogTransition.transitionType == TransitionType.none
-        ? showDialog<RateMyAppDialogButton>(context: context, builder: (context) => rateMyAppDialog)
+    RateMyAppDialogButton? clickedButton = await (dialogTransition
+                .transitionType ==
+            TransitionType.none
+        ? showDialog<RateMyAppDialogButton>(
+            context: context, builder: (context) => rateMyAppDialog)
         : showGeneralDialog<RateMyAppDialogButton>(
             context: context,
             barrierLabel: barrierLabel ?? '',
@@ -179,10 +188,11 @@ class RateMyApp {
     VoidCallback? onDismissed,
     bool? barrierDismissible,
     String? barrierLabel,
-    DialogTransition dialogTransition = const DialogTransition.def(),
+    DialogTransition dialogTransition = const DialogTransition(),
   }) async {
     ignoreNativeDialog ??= Platform.isAndroid;
-    if (!ignoreNativeDialog && ((await isNativeReviewDialogSupported) ?? false)) {
+    if (!ignoreNativeDialog &&
+        ((await isNativeReviewDialogSupported) ?? false)) {
       unawaited(callEvent(RateMyAppEventType.iOSRequestReview));
       await launchNativeReviewDialog();
       return;
@@ -191,11 +201,13 @@ class RateMyApp {
     assert(actionsBuilder != null, 'You should provide an actions builder.');
     unawaited(callEvent(RateMyAppEventType.starDialogOpen));
 
-    var starRateDialog = RateMyAppStarDialog(
+    RateMyAppStarDialog starRateDialog = RateMyAppStarDialog(
       this,
       title: title ?? 'Rate this app',
-      message: message ?? 'You like this app ? Then take a little bit of your time to leave a rating :',
-      contentBuilder: contentBuilder ?? ((context, defaultContent) => defaultContent),
+      message: message ??
+          'You like this app ? Then take a little bit of your time to leave a rating :',
+      contentBuilder:
+          contentBuilder ?? ((context, defaultContent) => defaultContent),
       actionsBuilder: actionsBuilder,
       dialogStyle: dialogStyle ??
           const DialogStyle(
@@ -207,7 +219,9 @@ class RateMyApp {
     );
 
     // Using [showDialog()] when [TransitionType.none] to get rid of the default fading animation of [showGeneralDialog()]
-    RateMyAppDialogButton? clickedButton = await (dialogTransition.transitionType == TransitionType.none
+    RateMyAppDialogButton? clickedButton = await (dialogTransition
+                .transitionType ==
+            TransitionType.none
         ? showDialog(context: context, builder: (context) => starRateDialog)
         : showGeneralDialog(
             context: context,
@@ -229,8 +243,8 @@ class RateMyApp {
 
   /// Launches the corresponding store.
   Future<LaunchStoreResult> launchStore() async {
-    int? result =
-        await _channel.invokeMethod<int>('launchStore', storeIdentifier == null ? null : {'appId': storeIdentifier});
+    int? result = await _channel.invokeMethod<int>('launchStore',
+        storeIdentifier == null ? null : {'appId': storeIdentifier});
     switch (result) {
       case 0:
         return LaunchStoreResult.storeOpened;
@@ -244,8 +258,8 @@ class RateMyApp {
   /// Calls the specified event.
   Future<void> callEvent(RateMyAppEventType eventType) async {
     bool saveSharedPreferences = false;
-    conditions
-        .forEach((condition) => saveSharedPreferences = condition.onEventOccurred(eventType) || saveSharedPreferences);
+    conditions.forEach((condition) => saveSharedPreferences =
+        condition.onEventOccurred(eventType) || saveSharedPreferences);
     if (saveSharedPreferences) {
       await save();
     }
@@ -277,31 +291,39 @@ class RateMyApp {
     switch (dialogTransition.transitionType) {
       case TransitionType.fade:
         return FadeTransition(
-          opacity: CurvedAnimation(curve: dialogTransition.curve, parent: animation),
+          opacity:
+              CurvedAnimation(curve: dialogTransition.curve, parent: animation),
           child: child,
         );
       case TransitionType.rotation:
         return RotationTransition(
-          turns: CurvedAnimation(curve: dialogTransition.curve, parent: animation),
+          turns:
+              CurvedAnimation(curve: dialogTransition.curve, parent: animation),
           child: child,
         );
       case TransitionType.scale:
         return ScaleTransition(
           alignment: dialogTransition.alignment ?? Alignment.center,
-          scale: CurvedAnimation(curve: dialogTransition.curve, parent: animation),
+          scale:
+              CurvedAnimation(curve: dialogTransition.curve, parent: animation),
           child: child,
         );
       case TransitionType.scaleAndFade:
         return FadeTransition(
-          opacity: CurvedAnimation(curve: dialogTransition.curve, parent: animation),
+          opacity:
+              CurvedAnimation(curve: dialogTransition.curve, parent: animation),
           child: ScaleTransition(
-            scale: CurvedAnimation(curve: dialogTransition.curve, parent: animation),
+            scale: CurvedAnimation(
+                curve: dialogTransition.curve, parent: animation),
             child: child,
           ),
         );
       case TransitionType.slide:
         return SlideTransition(
-          position: Tween<Offset>(begin: dialogTransition.startOffset ?? const Offset(1, 0), end: Offset.zero).animate(
+          position: Tween<Offset>(
+                  begin: dialogTransition.startOffset ?? const Offset(1, 0),
+                  end: Offset.zero)
+              .animate(
             CurvedAnimation(parent: animation, curve: dialogTransition.curve),
           ),
         );