+++ lib/src/dialogs.dart
@@ -1,23 +1,27 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_rating_bar/flutter_rating_bar.dart';
+import 'package:flutter_rating_bar/flutter_rating_bar.dart'
+    as flutter_rating_bar;
 import 'package:rate_my_app/rate_my_app.dart';
 import 'package:rate_my_app/src/core.dart';
 import 'package:rate_my_app/src/style.dart';
 
 /// A simple dialog button click listener.
-typedef RateMyAppDialogButtonClickListener = bool Function(RateMyAppDialogButton button);
+typedef RateMyAppDialogButtonClickListener = bool Function(
+    RateMyAppDialogButton button);
 
 /// Validates a state when called in a function.
 typedef Validator = bool Function();
 
 /// Allows to change the default dialog content.
-typedef DialogContentBuilder = Widget Function(BuildContext context, Widget defaultContent);
+typedef DialogContentBuilder = Widget Function(
+    BuildContext context, Widget defaultContent);
 
 /// Allows to dynamically build actions.
 typedef DialogActionsBuilder = List<Widget> Function(BuildContext context);
 
 /// Allows to dynamically build actions according to the specified rating.
-typedef StarDialogActionsBuilder = List<Widget> Function(BuildContext context, double? stars);
+typedef StarDialogActionsBuilder = List<Widget> Function(
+    BuildContext context, double? stars);
 
 /// The Android Rate my app dialog.
 class RateMyAppDialog extends StatelessWidget {
@@ -98,17 +102,20 @@ class RateMyAppDialog extends StatelessWidget {
         RateMyAppRateButton(
           rateMyApp,
           text: rateButton,
-          validator: () => listener == null || listener!(RateMyAppDialogButton.rate),
+          validator: () =>
+              listener == null || listener!(RateMyAppDialogButton.rate),
         ),
         RateMyAppLaterButton(
           rateMyApp,
           text: laterButton,
-          validator: () => listener == null || listener!(RateMyAppDialogButton.later),
+          validator: () =>
+              listener == null || listener!(RateMyAppDialogButton.later),
         ),
         RateMyAppNoButton(
           rateMyApp,
           text: noButton,
-          validator: () => listener == null || listener!(RateMyAppDialogButton.no),
+          validator: () =>
+              listener == null || listener!(RateMyAppDialogButton.no),
         ),
       ];
 }
@@ -151,7 +158,8 @@ class RateMyAppStarDialog extends StatefulWidget {
   State<StatefulWidget> createState() => _RateMyAppStarDialogState();
 
   /// Used when there is no onRatingChanged callback.
-  List<Widget> _defaultOnRatingChanged(BuildContext context, double? rating) => [
+  List<Widget> _defaultOnRatingChanged(BuildContext context, double? rating) =>
+      [
         RateMyAppRateButton(
           rateMyApp,
           text: 'RATE',
@@ -209,18 +217,20 @@ class _RateMyAppStarDialogState extends State<RateMyAppStarDialog> {
       content: widget.contentBuilder(context, content),
       contentPadding: widget.dialogStyle.contentPadding,
       shape: widget.dialogStyle.dialogShape,
-      actions: (widget.actionsBuilder ?? widget._defaultOnRatingChanged)(context, currentRating),
+      actions: (widget.actionsBuilder ?? widget._defaultOnRatingChanged)(
+          context, currentRating),
     );
   }
 
   /// Creates the rating bar.
-  RatingBar createRatingBar() {
+  flutter_rating_bar.RatingBar createRatingBar() {
     if (widget.starRatingOptions.itemBuilder == null) {
-      return RatingBar(
+      return flutter_rating_bar.RatingBar(
         onRatingUpdate: (rating) {
           setState(() => currentRating = rating);
         },
-        ratingWidget: widget.starRatingOptions.ratingWidget ?? createDefaultRatingWidget(),
+        ratingWidget: widget.starRatingOptions.ratingWidget ??
+            createDefaultRatingWidget(),
         initialRating: widget.starRatingOptions.initialRating,
         minRating: widget.starRatingOptions.minRating,
         allowHalfRating: widget.starRatingOptions.allowHalfRating,
@@ -233,7 +243,7 @@ class _RateMyAppStarDialogState extends State<RateMyAppStarDialog> {
         wrapAlignment: widget.starRatingOptions.wrapAlignment,
       );
     }
-    return RatingBar.builder(
+    return flutter_rating_bar.RatingBar.builder(
       onRatingUpdate: (rating) {
         setState(() => currentRating = rating);
       },
@@ -295,7 +305,8 @@ class _RateMyAppStarDialogState extends State<RateMyAppStarDialog> {
               ),
         empty: Icon(
           Icons.star_border,
-          color: widget.starRatingOptions.borderColor ?? widget.starRatingOptions.itemColor,
+          color: widget.starRatingOptions.borderColor ??
+              widget.starRatingOptions.itemColor,
           size: widget.starRatingOptions.itemSize,
         ),
       );